refactor: move pkg/mlsstore to pkg/mls/store
