+++ pkg/mls/store/config.go
@@ -1,4 +1,4 @@
-package mlsstore
+package store
 
 import (
 	"time"
